move get_dev_username
