+++ src/commcare_cloud/cli_utils.py
@@ -11,7 +11,6 @@ from six.moves import input, shlex_quote
 from commcare_cloud.colors import color_code, color_error
 
 from .environment.paths import ANSIBLE_DIR
-from .environment.main import get_environment
 
 
 def ask(message, strict=False, quiet=False):
@@ -34,31 +33,6 @@ def ask_option(message, options, acceptable_responses=None):
         r = input('Please enter one of {} :'.format(', '.join(options)))
     return r
 
-def get_dev_username(env_name):
-    from .commands.terraform.aws import (
-        print_help_message_about_the_commcare_cloud_default_username_env_var,
-    )
-    from .user_utils import get_default_username
-
-    environment = get_environment(env_name)
-    username = default_username = get_default_username()
-    while True:
-        if not username or default_username.is_guess:
-            username = input(f"Enter your SSH username ({default_username}): ")
-            if not username:
-                username = default_username
-        if username in environment.users_config.dev_users.present:
-            break
-        allowed_users = environment.users_config.dev_users.present
-        env_users = '\n  - '.join([''] + allowed_users)
-        puts(color_error(
-            f"Unauthorized user {username}.\n\n"
-            f"Please pass in one of the allowed ssh users:{env_users}"
-        ))
-        username = ""
-    if default_username.is_guess:
-        print_help_message_about_the_commcare_cloud_default_username_env_var(username)
-    return username
 
 def has_arg(unknown_args, short_form, long_form):
     """
